Criando calculadora com TDD - parte 1.1
Corrigindo um problema que pode acontecer relacionado ao tipo de dado
definido na operacao de divisao, o metodo nao pode retornar um valor
inteiro porque determinada operacao pode resultar em valor fracionado

Modifiquei os tipos das variaveis, os tipos dos parametros do metodo e
seu respectivo retorno para devolver valores do tipo double

Alem disso, na verificacao do teste deveDividirDoisValores, precisei
adicionar no metodo assertEquals um terceiro parametro para determinar
quantas casas decimais sera considerada no parametro que representa o
valor atual

+++ src/main/java/edu/testesunitarios/service/Calculadora.java
@@ -16,7 +16,7 @@ public class Calculadora {
 		return a * b;
 	}
 
-	public int dividir(int a, int b) throws NaoPodeDividirPorZeroException {
+	public double dividir(double a, double b) throws NaoPodeDividirPorZeroException {
 		if (b == 0) {
 			throw new NaoPodeDividirPorZeroException();
 		}
